Deleted dummy posts. Finished blog and releases pages
