+++ components/common/Menu.jsx
@@ -4,26 +4,63 @@ import { detectOverflow } from '@popperjs/core';
 
 import ReactDOM from 'react-dom';
 import styled from 'styled-components';
+import { mediaQueries } from '@/styles/breakpoints';
 
 const MenuWrap = styled.div`
+  min-width: 160px;
+  display: flex;
+  flex-direction: column;
+  z-index: 9996;
+  max-height: 400px;
+
+  @media ${mediaQueries.sm} {
+    position: fixed;
+    inset: auto auto 0 0 !important;
+    transform: none !important;
+    display: block;
+    /* height: calc(100% - 6rem); */
+    max-height: calc(100vh - 10rem - env(safe-area-inset-top) - env(safe-area-inset-bottom));
+    width: 100vw;
+    box-sizing: border-box;
+    z-index: 9999;
+    &:before {
+      background-color: #00000052;
+      content: "";
+      display: block;
+      height: 100vh;
+      left: 0;
+      position: fixed;
+      top: 0;
+      transition: opacity .4s cubic-bezier(.19,1,.22,1);
+      width: 100vw;
+      z-index: -1;
+      pointer-events: none;
+    }
+  }
+`
+const MenuInside = styled.div`
+  overflow-y: auto;
+  width: 100%;
   background: var(--background-tertiary-color);
   border: 1px solid var(--separator-color);;
   padding: 5px;
   border-radius: 10px;
-  min-width: 160px;
-  display: flex;
-  flex-direction: column;
   align-items: stretch;
   box-shadow: 0 8px 12px rgba(0, 0, 0, 0.15);
-  z-index: 9996;
-  max-height: 400px;
-  overflow-y: auto;
+  box-sizing: border-box;
+
+  @media ${mediaQueries.sm} {
+    border: 0;
+    max-height: calc(100vh - 10rem - env(safe-area-inset-top) - env(safe-area-inset-bottom));
+    border-radius: 10px 10px 0 0;
+  }
 `
 
 const MenuItemWrap = styled.button`
   display: flex;
   gap: 10px;
   align-items: center;
+  width: 100%;
   padding: 6px 10px;
   border-radius: 5px;
   background: transparent;
@@ -31,15 +68,41 @@ const MenuItemWrap = styled.button`
   color: inherit;
   cursor: pointer;
   text-align: left;
-  &:hover {
+  box-sizing: border-box;
+  &:hover, &:focus {
     background: var(--fill-gray-quaternary);
   }
   &:active {
     background: var(--fill-gray-tertiary);
   }
+  @media ${mediaQueries.sm} {
+    padding: 14px 15px;
+    border-top: 1px solid --var(--separator-color);
+    font-size: 14px;
+    &:has(svg) {
+      flex-direction: row-reverse;
+      justify-content: space-between;      
+    }
+  }
+  
   svg {
     width: 16px;
     height: 16px;
+  }
+`
+const MenuCancel = styled.div`
+  display: none;
+  position: sticky;
+  bottom: 0;
+  padding-bottom: 5px;
+  margin-top: -5px;
+  transform: translateY(5px);
+  background: var(--background-tertiary-color);
+  ${MenuItemWrap} {
+    justify-content: center;
+  }
+  @media ${mediaQueries.sm} {
+    display: block;
   }
 `
 
@@ -50,13 +113,13 @@ const MenuDivider = styled.hr`
   border-left: 1px solid var(--separator-color);;
 `
 
-const Menu = ({ trigger, children, ...props }) => {
+const Menu = ({ trigger, children, placement = 'bottom-end', ...props }) => {
   const [isOpen, setIsOpen] = useState(false);
   const [triggerEl, setTriggerEl] = useState(null);
   const [popperEl, setPopperEl] = useState(null);
 
   const { styles, attributes } = usePopper(triggerEl, popperEl, {
-    placement: 'bottom-end',
+    placement,
     strategy: 'fixed',
     modifiers: [
     {
@@ -100,7 +163,13 @@ const Menu = ({ trigger, children, ...props }) => {
 
   const menu = (
     <MenuWrap ref={setPopperEl} style={styles.popper} {...attributes.popper} {...props}>
-      {clonedChildren}
+      <MenuInside>
+        {clonedChildren}
+        <MenuCancel>
+          <MenuDivider />
+          <MenuItemWrap onClick={() => setIsOpen(false)}>Cancel</MenuItemWrap>
+        </MenuCancel>
+      </MenuInside>
     </MenuWrap>
   )
