--- data/blog/2023/12/redesigned-apple-tv-app-elevates-the-viewing-experience.md
@@ -1,92 +1,0 @@
----
-title: Redesigned Apple TV app elevates the viewing experience
-headline: Redesigned Apple TV app elevates the viewing experience
-subhead: The intuitive new interface lets users more easily browse and watch Apple Original series and films, as well as live sports, movies, and more
-description: Redesigned Apple TV app elevates the viewing experience
-date: December 11, 2023
-category: update
-author: thecoolwinter
-image: https://www.apple.com/newsroom/images/2023/12/redesigned-apple-tv-app-simplifies-the-viewing-experience/article/Apple-TV-app-home-screen_big.jpg.large_2x.jpg
----
-
-<figure>
-  <img class="full-width" src="https://www.apple.com/newsroom/images/2023/12/redesigned-apple-tv-app-simplifies-the-viewing-experience/article/Apple-TV-app-home-screen_big.jpg.large_2x.jpg" alt="An image description">
-  <figcaption>Available today, the redesigned Apple TV app allows users to easily access all of Apple’s award-winning programming across their favorite devices.</figcaption>
-</figure>
-
-Today Apple unveiled a redesigned Apple TV app that makes it even easier for users to watch Apple’s award-winning original series and films, as well as live sports, movies, and television shows across their favorite Apple devices, smart TVs, streaming sticks, gaming consoles, and more. The simplified interface’s new sidebar allows users to quickly navigate the app, providing easy access to Apple TV+, home of Apple Original series and films; MLS Season Pass, home of Major League Soccer; Store, where users can buy and rent popular movies; and shortcuts to channels and apps viewers already have, such as Disney+, Paramount+, and Max.
-
-<blockquote>
-The redesigned Apple TV app makes it easier than ever for users to watch the shows, movies, and sports they love
-</blockquote>
-
-“The redesigned Apple TV app makes it easier than ever for users to watch the shows, movies, and sports they love through an intuitive interface that brings content to the forefront,” said Eddy Cue, Apple’s senior vice president of Services. “With so much available to watch, our aim is to ensure users always have their favorites at their fingertips.”
-
-## Fruit
-
-1. Apple
-2. Pear
-3. Plumb
-4. Apricot
-5. Mango
-6. Orange
-7. Banana
-8. Grape
-9. Cherry
-10. Strawberry
-11. Blueberry
-12. Rasberry
-13. Blackberry
-14. Watermelon
-15. Cantelope
-16. Honeydew
-
-## Test
-
-Test test test
-
-- Here is one point
-- And this is another point
-- This is a whole different point
-
-| Package             | Source                                                   | Author                                        |
-| :------------------ | :------------------------------------------------------- | :-------------------------------------------- |
-| `TextStory`         | [GitHub](https://github.com/ChimeHQ/TextStory)           | [Matt Massicotte](https://twitter.com/mattie) |
-| `swift-collections` | [GitHub](https://github.com/apple/swift-collections.git) | [Apple](https://github.com/apple)             |
-
-### Test
-
-Test test test
-
-#### Test
-
-The new sidebar navigation also introduces Home, a unified guide for all the shows, movies, and sports viewers love. Within Home, the Channels & Apps section allows users to browse each of their subscribed channels or connected apps in depth. And collections — including New Shows & Movies, Top Charts, Trending, and For You — bring forward the best recommendations for viewers to enjoy across what’s new, popular, and tailored just for them.
-
-<figure>
-  <img src="https://www.apple.com/newsroom/images/2023/12/redesigned-apple-tv-app-simplifies-the-viewing-experience/article/Apple-TV-app-iOS-app-home-screen_inline.jpg.large_2x.jpg" alt="An image description">
-  <figcaption>The redesigned navigation introduces Home, a unified guide for all the shows, movies, and sports users love.</figcaption>
-</figure>
-
-On `living room` devices, the sidebar will also feature profiles, allowing households to quickly switch between users for better personalization in Up Next and content recommendations across the app.
-
-<figure>
-  <img class="wide" src="https://www.apple.com/newsroom/images/2023/12/redesigned-apple-tv-app-simplifies-the-viewing-experience/article/Apple-TV-app-profiles_big.jpg.large_2x.jpg" alt="An image description">
-  <figcaption>Profiles allow households to quickly switch between users on living room devices.
-</figcaption>
-</figure>
-
-The updated Store tab in the Apple TV app unifies movies and TV shows into one offering, so users can more easily access everything available to buy or rent, and a new Add Channels & Apps shelf where users can explore popular streaming services.
-
-```swift
-func message(_ name: String) {
-  return "Hello, \(name)"
-}
-```
-
-<figure>
-  <img class="wide" src="https://www.apple.com/newsroom/images/2023/12/redesigned-apple-tv-app-simplifies-the-viewing-experience/article/Apple-TV-app-iPadOS-App-Store_big.jpg.large_2x.jpg" alt="An image description">
-  <figcaption>In the Apple TV app, the updated Store tab unifies movies and TV shows into one offering.
-</figcaption>
-</figure>
-
-Also starting today, the iTunes Store app on iPhone and iPad, and the iTunes Movies and iTunes TV Shows apps on Apple TV 4K and Apple TV HD, will redirect users to the Apple TV app, where they can find their existing purchases and watch all of their favorite TV shows and movies in one place.
