+++ data/blog/2024/01/why-were-building-codeedit.md
@@ -1,0 +1,87 @@
+---
+title: "Why We're Building CodeEdit: The Next Revolution of Code Editing on the Mac"
+headline: "Why We're Building CodeEdit: The Next Revolution of Code Editing on the Mac"
+subhead: It might sound crazy, but we set out to build an editor native to macOS written in Swift, by the community, for the community. Completely open source and free forever.
+description: Redesigned Apple TV app elevates the viewing experience
+date: December 26, 2023
+category: update
+author: austincondiff
+image: https://github.com/CodeEditApp/codeedit.app/assets/806104/31579fc0-11ce-4568-b9cb-84ab89086824
+---
+
+<img src="https://github.com/CodeEditApp/codeedit.app/assets/806104/31579fc0-11ce-4568-b9cb-84ab89086824" alt="Community building software" class="wide" />
+
+Most code editors today are built on cross-platform frameworks and technologies that often fall short in fully leveraging the capabilities of macOS and taking full advantage of system resources. There's a significant gap in the market for a truly native code editing experience on macOS.
+
+CodeEdit is our answer to this gap. It's not just an application; it's a concept turned reality, born from the question, "What if there was an editor that could harness the full potential of the Mac while offering a native experience?" 
+
+While Xcode does provide a native experience, it's tailored specifically for Apple platform development, leaving a void for those working on projects outside of this ecosystem or in support of it.
+
+Our goal is to offer a familiar, intuitive editing experience to developers working on projects outside Apple's ecosystem.
+
+## Community-Driven, Open Source, and Free
+
+At the core of CodeEdit lies our commitment to the community, as we believe in the power of community-led projects and welcome contributions from everyone. This approach defines not only the development process but also the essence of the application, setting new standards in software development. We are dedicated to keeping CodeEdit open source and free forever, reflecting its strong roots in the community.
+
+### Rapid Development and Iteration
+
+In a community-driven project, the development cycle is incredibly dynamic. We can often times build, iterate, and refine features much faster than in a traditional setup. This agility allows us to quickly adapt to new ideas and feedback, ensuring that CodeEdit evolves in a way that consistently meets the needs of its users.
+
+### Independence and Diversity of Thought
+
+Unlike projects dictated by company directors or a specific set of stakeholders, CodeEdit benefits from the diverse range of ideas and perspectives of the community. This independence means we're not limited by a single company's vision or agenda. Instead, we can explore creative solutions and innovative features that a more corporate-driven project might overlook.
+
+### Amplified Feedback and Collaboration
+
+With a broader community of users and contributors, we have access to a more extensive feedback pool. This diverse input is crucial in identifying and fixing issues quickly, enhancing usability, and ensuring that the editor meets a wide range of developer needs.
+
+### Made by the Community, For the Community
+
+Every major change, feature, or improvement in CodeEdit is subject to community review and approval. This approach ensures that the developments in the editor are not only technically sound but also genuinely desired by its user base. It's a process that values every voice in the community, fostering a sense of ownership and engagement among all contributors.
+
+## The Benefits of a Native macOS Code Editor
+
+### Unmatched Performance
+
+By utilizing macOS's native capabilities, CodeEdit offers a level of performance that cross-platform editors can't match. This means faster load times, smoother operation, and better resource management.
+
+### A Familiar, Intuitive Interface
+
+Staying true to Apple's human interface guidelines, CodeEdit provides an environment that feels intuitive and familiar to macOS users. It's designed to feel like an extension of macOS, not just another third-party application.
+
+The intention is to help developers stay focused and connected to the code being written by building an application where the UI simply melts into the background and allowing developers to interact with their code in new ways. The best UI is invisible and we think that conforming to the design patterns Apple has put into place, we can achieve just that.
+
+### Scalability and Flexibility
+
+From lightweight text editing to handling complex development tasks, CodeEdit is scalable. It combines the simplicity of TextEdit with the power of Xcode, ensuring a flexible environment that adapts to your needs.
+
+## Our Story: How it all began
+
+In the world of software development, necessity is often the mother of invention. This is precisely the case with CodeEdit, a project that began not in a boardroom, but from a developer's personal challenge and a vision for something better. Here's how this groundbreaking idea came to life.
+
+### A Personal Project Born From Frustration with Existing Tools
+
+My journey with CodeEdit started as a personal quest to enhance my existing editor. I began by customizing VS Code, trying to give it a macOS-native feel. This involved writing custom CSS and applying it to VS Code after each update. However, this was far from a perfect solution and by no means scalable.
+
+At around the same time, the limitations of VS Code were becoming increasingly apparent. I would encounter hangs, crashes, and various performance issues that interrupted my workflow. These frustrations led me to a pivotal question: If I were to design a code editor from scratch, specifically for macOS, what would it look like?
+
+### From Mockups to Movement
+
+This question didn't remain hypothetical for long. I started by experimenting with a few designs. These mockups rapidly evolved into a comprehensive user flow comprising 20-30 designs. The more I designed, the more the vision of a native macOS code editor took shape.
+
+I couldn't keep these ideas to myself. I shared my designs on social media, sparking a conversation about the need for a more general-purpose code editor for macOS. The underlying thought was simple yet powerful: If Microsoft could create something as successful as VS Code, why couldn't Apple, or the community, build a code editor native to macOS?
+
+Despite reaching out to several Xcode engineers at Apple and learning that direct assistance wasn't forthcoming, I was encouraged to embark on an open-source journey. This challenge was daunting, especially as a JavaScript developer new to Swift. Nevertheless, I was determined to give it a shot.
+
+I started with a simple "Hello World" application resembling a code editor's layout. Accompanied by a readme explaining our goals and mission, this humble beginning was the seed that would grow into something much larger.
+
+When I shared the repository on social media, the response was overwhelming. The project resonated with the developer community, amassing over 5,000 stars in just a few weeks. Contributions started flowing in, with developers from around the world submitting pull requests.
+
+Now, CodeEdit stands as one of the most starred repositories written in Swift on GitHub. Our journey from a set of mockups to a community-driven open-source project exemplifies the power of collaboration and shared vision.
+
+## The Future is Bright
+
+As CodeEdit continues to grow and evolve, we're more committed than ever to our mission of keeping it open source, free forever. We're not just building a code editor; we're creating a community, a movement. 
+
+Whether you're a developer, a designer, or simply someone who believes in the power of open-source software, together we can make something truly special.
+
