+++ data/blog/starter-post.md
@@ -1,0 +1,30 @@
+---
+title: Starter post title
+description: Description for SEO
+headline: This is a longer starter post title
+subhead: Longer description to show underneath headline at the top of the post
+date: 2024-01-20
+category: update
+author: austincondiff
+image: https://github.com/CodeEditApp/codeedit.app/assets/806104/6552453d-c31c-48f3-ae0f-862db07ada9c
+draft: true
+---
+
+![image](https://github.com/CodeEditApp/codeedit.app/assets/806104/6552453d-c31c-48f3-ae0f-862db07ada9c)
+
+This is some body text.
+
+## Heading 2
+
+### Heading 3
+
+Here is some more body text. Check out this inline `code`.
+
+> [!NOTE]
+> This is a note.
+
+```javascript
+alert("Look, it's a codeblock")
+```
+
+For more examples, check out the blog [demo](/blog/demo).
