--- data/blog/test01.md
@@ -1,41 +1,0 @@
----
-title: Test 1 shorter title to display on post cards
-description: Description for SEO
-headline: Test 1 longer title to show at the top of the post
-subhead: Longer description to show underneath headline at the top of the post
-date: 2023-11-01
-category: update
-author: austincondiff
-image: https://www.apple.com/newsroom/images/product/ipad/standard/Apple-iPad-Air-feature-Belinda-Kou.jpg.landing-gradient-large_2x.jpg
----
-
-Intro text goes here.
-
-## Section
-
-Text about this section goes here `const`.
-
-```
-Code block with no language. Code block with no language. Code block with no language.
-
-So sad. :(
-```
-
-```js
-function () {
-  const message = "Code block with a language.";
-
-  return message;
-}
-```
-
-## Section
-
-Text about this section goes here.
-
-> [!IMPORTANT]
-> This is important
-
-## Section
-
-Text about this section goes here.
