--- data/blog/test02.md
@@ -1,24 +1,0 @@
----
-title: Test 2 shorter title to display on post cards
-description: Description for SEO
-headline: Test 2 longer title to show at the top of the post
-subhead: Longer description to show underneath headline at the top of the post
-date: 2023-12-02
-category: update
-author: thecoolwinter
-image: https://www.apple.com/newsroom/images/product/iphone/lifestyle/Apple-Shot-on-iPhone-macro-Jirasak-Panpiansin_big.jpg.slideshow-xlarge_2x.jpg
----
-
-Intro text goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
