--- data/blog/test04.md
@@ -1,24 +1,0 @@
----
-title: Test 4 shorter title to display on post cards
-description: Description for SEO
-headline: Test 4 longer title to show at the top of the post
-subhead: Longer description to show underneath headline at the top of the post
-date: 2023-12-03
-category: update
-author: austincondiff
-image: https://www.apple.com/newsroom/images/product/app-store/Apple-App-Store-developers-job-growth-hero_big.jpg.large_2x.jpg
----
-
-Intro text goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
