--- data/blog/test07.md
@@ -1,24 +1,0 @@
----
-title: Test 7 shorter title to display on post cards
-description: Description for SEO
-headline: Test 7 longer title to show at the top of the post
-subhead: Longer description to show underneath headline at the top of the post
-date: 2023-12-03
-category: update
-author: austincondiff
-image: https://www.apple.com/newsroom/images/passions/education/Apple_gallaudet-university_10072020_Full-Bleed-Image.jpg.large_2x.jpg
----
-
-Intro text goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
