--- data/blog/test08.md
@@ -1,24 +1,0 @@
----
-title: Test 8 shorter title to display on post cards
-description: Description for SEO
-headline: Test 8 longer title to show at the top of the post
-subhead: Longer description to show underneath headline at the top of the post
-date: 2023-12-03
-category: update
-author: austincondiff
-image: https://www.apple.com/newsroom/images/passions/education/Apple_community-education-initiative-coding-academy-portrice-warren-home-office-birmingham-alabama_08272020_big.jpg.large_2x.jpg
----
-
-Intro text goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
-
-## Section
-
-Text about this section goes here.
