+++ data/macOS-versions.json
@@ -1,0 +1,22 @@
+{
+  "10.0": "Cheetah",
+  "10.1": "Puma",
+  "10.2": "Jaguar",
+  "10.3": "Panther",
+  "10.4": "Tiger",
+  "10.5": "Leopard",
+  "10.6": "Snow Leopard",
+  "10.7": "Lion",
+  "10.8": "Mountain Lion",
+  "10.9": "Mavericks",
+  "10.10": "Yosemite",
+  "10.11": "El Capitan",
+  "10.12": "Sierra",
+  "10.13": "High Sierra",
+  "10.14": "Mojave",
+  "10.15": "Catalina",
+  "11": "Big Sur",
+  "12": "Monterey",
+  "13": "Ventura",
+  "14": "Sonoma"
+}
